Refactoring - rename package actions to panelstates - adjust some comments
