+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/common/control/FeaturePanel.java
@@ -48,6 +48,12 @@ public abstract class FeaturePanel<C extends FeaturePanel<C, S>, S extends Featu
         this(new SimpleBooleanProperty(false));
     }
 
+    /**
+     * Closes the currently opened details panel.
+     * If no details panel has been opened nothing happens
+     */
+    public abstract void closeDetailsPanel();
+
     public boolean isInitialized() {
         return this.initialized.get();
     }
@@ -59,10 +65,4 @@ public abstract class FeaturePanel<C extends FeaturePanel<C, S>, S extends Featu
     public void setInitialized(boolean initialized) {
         this.initialized.set(initialized);
     }
-
-    /**
-     * Closes the currently opened details panel.
-     * If no details panel has been opened nothing happens
-     */
-    public abstract void closeDetailsPanel();
 }
