+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/common/panelstates/OpenDetailsPanel.java
@@ -1,4 +1,4 @@
-package org.phoenicis.javafx.components.common.actions;
+package org.phoenicis.javafx.components.common.panelstates;
 
 /**
  * An interface that must be implemented by all indicator objects that specify a shown/opened details panel and its
