+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/container/skin/ContainersFeaturePanelSkin.java
@@ -15,15 +15,15 @@ import org.phoenicis.containers.dto.ContainerDTO;
 import org.phoenicis.engines.EngineSetting;
 import org.phoenicis.javafx.collections.ConcatenatedList;
 import org.phoenicis.javafx.collections.MappedList;
-import org.phoenicis.javafx.components.common.actions.None;
-import org.phoenicis.javafx.components.common.actions.OpenDetailsPanel;
+import org.phoenicis.javafx.components.common.panelstates.None;
+import org.phoenicis.javafx.components.common.panelstates.OpenDetailsPanel;
 import org.phoenicis.javafx.components.common.control.DetailsPanel;
 import org.phoenicis.javafx.components.common.control.SidebarBase;
 import org.phoenicis.javafx.components.common.skin.FeaturePanelSkin;
 import org.phoenicis.javafx.components.common.widgets.control.CombinedListWidget;
 import org.phoenicis.javafx.components.common.widgets.utils.ListWidgetElement;
 import org.phoenicis.javafx.components.common.widgets.utils.ListWidgetType;
-import org.phoenicis.javafx.components.container.actions.ContainerInformation;
+import org.phoenicis.javafx.components.container.panelstates.ContainerInformation;
 import org.phoenicis.javafx.components.container.control.ContainerInformationPanel;
 import org.phoenicis.javafx.components.container.control.ContainerSidebar;
 import org.phoenicis.javafx.components.container.control.ContainersFeaturePanel;
