+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/installation/skin/InstallationsFeaturePanelSkin.java
@@ -10,15 +10,15 @@ import javafx.collections.transformation.SortedList;
 import javafx.scene.Node;
 import org.phoenicis.javafx.collections.ConcatenatedList;
 import org.phoenicis.javafx.collections.MappedList;
-import org.phoenicis.javafx.components.common.actions.None;
-import org.phoenicis.javafx.components.common.actions.OpenDetailsPanel;
+import org.phoenicis.javafx.components.common.panelstates.None;
+import org.phoenicis.javafx.components.common.panelstates.OpenDetailsPanel;
 import org.phoenicis.javafx.components.common.control.DetailsPanel;
 import org.phoenicis.javafx.components.common.control.SidebarBase;
 import org.phoenicis.javafx.components.common.skin.FeaturePanelSkin;
 import org.phoenicis.javafx.components.common.widgets.control.CombinedListWidget;
 import org.phoenicis.javafx.components.common.widgets.utils.ListWidgetElement;
 import org.phoenicis.javafx.components.common.widgets.utils.ListWidgetType;
-import org.phoenicis.javafx.components.installation.actions.Installation;
+import org.phoenicis.javafx.components.installation.panelstates.Installation;
 import org.phoenicis.javafx.components.installation.control.InstallationSidebar;
 import org.phoenicis.javafx.components.installation.control.InstallationsFeaturePanel;
 import org.phoenicis.javafx.settings.JavaFxSettingsManager;
