--- phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/library/actions/ShortcutCreation.java
@@ -1,9 +1,0 @@
-package org.phoenicis.javafx.components.library.actions;
-
-import org.phoenicis.javafx.components.common.actions.OpenDetailsPanel;
-
-/**
- * Indicator to open the "Shortcut Creation" details panel
- */
-public class ShortcutCreation implements OpenDetailsPanel {
-}
