+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/library/panelstates/ShortcutCreation.java
@@ -1,0 +1,9 @@
+package org.phoenicis.javafx.components.library.panelstates;
+
+import org.phoenicis.javafx.components.common.panelstates.OpenDetailsPanel;
+
+/**
+ * Indicator to open the "Shortcut Creation" details panel
+ */
+public class ShortcutCreation implements OpenDetailsPanel {
+}
