+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/library/skin/LibraryFeaturePanelSkin.java
@@ -14,17 +14,17 @@ import javafx.scene.control.TabPane;
 import javafx.scene.input.MouseEvent;
 import org.phoenicis.javafx.collections.ConcatenatedList;
 import org.phoenicis.javafx.collections.MappedList;
-import org.phoenicis.javafx.components.common.actions.None;
-import org.phoenicis.javafx.components.common.actions.OpenDetailsPanel;
+import org.phoenicis.javafx.components.common.panelstates.None;
+import org.phoenicis.javafx.components.common.panelstates.OpenDetailsPanel;
 import org.phoenicis.javafx.components.common.control.DetailsPanel;
 import org.phoenicis.javafx.components.common.control.SidebarBase;
 import org.phoenicis.javafx.components.common.skin.FeaturePanelSkin;
 import org.phoenicis.javafx.components.common.widgets.control.CombinedListWidget;
 import org.phoenicis.javafx.components.common.widgets.utils.ListWidgetElement;
 import org.phoenicis.javafx.components.common.widgets.utils.ListWidgetType;
-import org.phoenicis.javafx.components.library.actions.ShortcutCreation;
-import org.phoenicis.javafx.components.library.actions.ShortcutInformation;
-import org.phoenicis.javafx.components.library.actions.ShortcutEditing;
+import org.phoenicis.javafx.components.library.panelstates.ShortcutCreation;
+import org.phoenicis.javafx.components.library.panelstates.ShortcutInformation;
+import org.phoenicis.javafx.components.library.panelstates.ShortcutEditing;
 import org.phoenicis.javafx.components.library.control.*;
 import org.phoenicis.javafx.settings.JavaFxSettingsManager;
 import org.phoenicis.javafx.utils.ObjectBindings;
